pyproject.toml
[project]
name = "dokploy-seed"
version = "0.1.0"
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]
requires-python = ">=3.13,<3.14"

dependencies = [
    "httpx>=0.28.1,<1",
    "python-decouple>=3.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.5",
]

test = [
    "coverage<8.0.0,>=7.6.1",
    "hypothesis[cli]<7.0.0,>=6.112.1",
    "pytest<9.0.0,>=8.3.3",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "pytest-cov>=6.1.1",
    "pytest-datafiles<4.0.0,>=3.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist<4.0.0,>=3.6.1",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
addopts = "-m 'not e2e'"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "benchmark: marks performance benchmark tests",
    "property: marks property-based tests (Hypothesis)"
]

[project.scripts]
dps = "dokploy_seed:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
